Add tests for field interceptors
