+++ ballerina-tests/tests/resources/documents/interceptor_execution_order.graphql
@@ -1,3 +1,4 @@
 {
-    quote
+    quote,
+    status
 }
