+++ ballerina-tests/tests/resources/documents/interceptors_returning_invalid_value.graphql
@@ -1,3 +1,4 @@
 {
     name
+    age
 }
