+++ ballerina-tests/tests/resources/documents/interceptors_with_destructive_modification1.graphql
@@ -2,5 +2,10 @@
     students {
         id,
         name
+    },
+    teachers {
+        id,
+        name,
+        subject
     }
 }
