+++ ballerina-tests/tests/resources/documents/interceptors_with_destructive_modification2.graphql
@@ -2,5 +2,9 @@
     students {
         id,
         name
+    },
+    teachers {
+        id,
+        name
     }
 }
