+++ ballerina-tests/tests/resources/documents/interceptors_with_mutation.graphql
@@ -1,5 +1,9 @@
 mutation {
     setName(name: "Heisenberg") {
         name
+    },
+    setAge(age: 14) {
+        name
+        age
     }
 }
