+++ ballerina-tests/tests/resources/documents/interceptors_with_null_value2.graphql
@@ -1,3 +1,4 @@
 {
-    name
+    name,
+    age
 }
