+++ ballerina-tests/tests/resources/documents/interceptors_with_records.graphql
@@ -5,5 +5,8 @@
             number,
             street
         }
+    },
+    contact {
+        number
     }
 }
