+++ ballerina-tests/tests/resources/documents/interceptors_with_service_object.graphql
@@ -3,5 +3,9 @@
         id,
         name,
         subject
+    },
+    student{
+        id,
+        name
     }
 }
