+++ ballerina-tests/tests/resources/documents/interceptors_with_subscription_return_records.graphql
@@ -1,0 +1,13 @@
+subscription A{ 
+    books {
+        name 
+        author  
+    } 
+}
+
+subscription B{ 
+    newBooks {
+        name 
+        author  
+    } 
+}
