+++ ballerina-tests/tests/resources/documents/interceptors_with_subscription_return_union_type.graphql
@@ -1,0 +1,22 @@
+subscription unionTypes1 {
+    multipleValues1 {
+        ...StudentFragment
+        ...TeacherFragment
+    }
+}
+
+subscription unionTypes2 {
+    multipleValues2 {
+        ...TeacherFragment
+    }
+}
+
+fragment StudentFragment on StudentService {
+    id
+    name
+}
+
+fragment TeacherFragment on TeacherService {
+    name
+    subject
+}
