+++ ballerina-tests/tests/resources/documents/interceptors_with_table.graphql
@@ -1,5 +1,8 @@
 {
     employees {
         name
+    },
+    oldEmployees {
+        name
     }
 }
