+++ ballerina-tests/tests/resources/documents/interceptors_with_union.graphql
@@ -1,8 +1,15 @@
 {
-    profile(id: 4) {
+    profile1(id: 4) {
         ...on StudentService {
             id,
             name
+        }
+    },
+    profile2(id: 4) {
+        ...on TeacherService {
+            id,
+            name,
+            subject
         }
     }
 }
