+++ ballerina-tests/tests/resources/documents/resource_interceptors.graphql
@@ -1,0 +1,11 @@
+query {
+    customer {
+        id
+        name
+        address {
+            number
+            street
+            city
+        }
+    }
+}
