+++ ballerina-tests/tests/resources/expected_results/duplicate_interceptors.json
@@ -1,5 +1,5 @@
 {
     "data":{
-        "age": 26
+        "age": 29
     }
 }
