+++ ballerina-tests/tests/resources/expected_results/execute_same_interceptor_multiple_times.json
@@ -1,5 +1,5 @@
 {
     "data":{
-        "age": 26
+        "age": 29
     }
 }
