+++ ballerina-tests/tests/resources/expected_results/interceptor_execution_order.json
@@ -1,5 +1,6 @@
 {
     "data": {
-        "quote": "Ballerina is an open-source programming language."
+        "quote": "Ballerina is an open-source programming language.",
+        "status": "Ballerina is a powerful general-purpose programming language."
     }
 }
