+++ ballerina-tests/tests/resources/expected_results/interceptors.json
@@ -1,5 +1,5 @@
 {
     "data":{
-        "enemy": "Tom Marvolo Riddle - voldemort"
+        "enemy": "Harry Potter and the Chamber of Secrets --> Tom Marvolo Riddle - voldemort"
     }
 }
