+++ ballerina-tests/tests/resources/expected_results/interceptors_with_destructive_modification1.json
@@ -5,6 +5,13 @@
                 "id": 3,
                 "name": "Minerva McGonagall"
             }
+        ],
+        "teachers": [
+            {
+                "id": 46,
+                "name": "Sybill Trelawney",
+                "subject": "Divination"
+            }
         ]
     }
 }
