+++ ballerina-tests/tests/resources/expected_results/interceptors_with_destructive_modification2.json
@@ -1,5 +1,6 @@
 {
     "data": {
-        "students": ["Ballerina","GraphQL"]
+        "students": ["Ballerina","GraphQL"],
+        "teachers": ["Hello","World!"]
     }
 }
