+++ ballerina-tests/tests/resources/expected_results/interceptors_with_invalid_destructive_modification1.json
@@ -1,6 +1,30 @@
 {
     "errors": [
         {
+            "message":"Invalid return type in Interceptor \"InvalidInterceptor9\". Expected type Int!",
+            "locations":[
+                {
+                    "line":2,
+                    "column":5
+                }
+            ],
+            "path":[
+                "age"
+            ]
+        },
+        {
+            "message":"Invalid return type in Interceptor \"InvalidInterceptor8\". Expected type Int!",
+            "locations":[
+                {
+                    "line":2,
+                    "column":5
+                }
+            ],
+            "path":[
+                "age"
+            ]
+        },
+        {
             "message": "Invalid return type in Interceptor \"InvalidInterceptor2\". Expected type Int!",
             "locations": [
                 {
