+++ ballerina-tests/tests/resources/expected_results/interceptors_with_invalid_destructive_modification2.json
@@ -1,6 +1,18 @@
 {
     "errors": [
         {
+            "message":"Invalid return type in Interceptor \"InvalidInterceptor8\". Expected type [String!]!",
+            "locations":[
+                {
+                    "line":2,
+                    "column":5
+                }
+            ],
+            "path": [
+                "friends"
+            ]
+        },
+        {
             "message": "Invalid return type in Interceptor \"InvalidInterceptor4\". Expected type [String!]!",
             "locations": [
                 {
