+++ ballerina-tests/tests/resources/expected_results/interceptors_with_invalid_destructive_modification3.json
@@ -1,6 +1,18 @@
 {
     "errors": [
         {
+            "message":"Invalid return type in Interceptor \"InvalidInterceptor9\". Expected type Person!",
+            "locations":[
+                {
+                    "line":2,
+                    "column":5
+                }
+            ],
+            "path":[
+                "person"
+            ]
+        },
+        {
             "message": "Invalid return type in Interceptor \"InvalidInterceptor6\". Expected type Person!",
             "locations": [
                 {
