+++ ballerina-tests/tests/resources/expected_results/interceptors_with_map.json
@@ -5,6 +5,12 @@
             "frontend": "Flutter",
             "data": "Ballerina",
             "native": "C#"
+        },
+        "updatedLanguages":{
+            "backend":"PHP",
+            "frontend":"JavaScript",
+            "data":"Python",
+            "native":"C#"
         }
     }
 }
