+++ ballerina-tests/tests/resources/expected_results/interceptors_with_mutation.json
@@ -2,6 +2,10 @@
     "data": {
         "setName": {
             "name": "Albus Percival Wulfric Brian Dumbledore"
+        },
+        "setAge": {
+            "name": "Albert",
+            "age": 53
         }
     }
 }
