+++ ballerina-tests/tests/resources/expected_results/interceptors_with_null_value2.json
@@ -1,5 +1,6 @@
 {
     "data": {
-        "name": null
+        "name": null,
+        "age": null
     }
 }
