+++ ballerina-tests/tests/resources/expected_results/interceptors_with_records.json
@@ -6,6 +6,9 @@
                 "number": "103",
                 "street": "Mould-on-the-Wold"
             }
+        },
+        "contact": {
+            "number": "+87654321"
         }
     }
 }
