+++ ballerina-tests/tests/resources/expected_results/interceptors_with_service_object.json
@@ -4,6 +4,10 @@
             "id": 3,
             "name": "Minerva McGonagall",
             "subject": "Transfiguration"
+        },
+        "student": {
+            "id": 45,
+            "name": "Ron Weasley"
         }
     }
 }
