+++ ballerina-tests/tests/resources/expected_results/interceptors_with_table.json
@@ -10,6 +10,17 @@
             {
                 "name": "Eng. Johnny Roe"
             }
+        ],
+        "oldEmployees":[
+            {
+                "name":"John"
+            },
+            {
+                "name":"Jane"
+            },
+            {
+                "name":"Johnny"
+            }
         ]
     }
 }
