+++ ballerina-tests/tests/resources/expected_results/interceptors_with_union.json
@@ -1,8 +1,13 @@
 {
     "data": {
-        "profile": {
+        "profile1": {
             "id": 3,
             "name": "Minerva McGonagall"
+        },
+        "profile2": {
+            "id": 4,
+            "name": "Minerva McGonagall",
+            "subject": "Black Magic"
         }
     }
 }
